04-GCP/requirements.txt
# Core Google Cloud dependencies
google-cloud-compute>=1.14.0
google-cloud-storage>=2.10.0
google-cloud-resourcemanager>=1.10.0

# Authentication and core functionality
google-auth>=2.17.0
google-api-core>=2.11.0

# Additional dependencies for enhanced functionality
google-cloud-iam>=2.12.0 